+++ src/GSM610/lpc.c
@@ -204,7 +204,7 @@ static void Fast_Autocorrelation(int16_t *s, /* [0..159] IN/OUT  */
     scale = MAX_LONGWORD / f_L_ACF[0];
 
     for (k = 0; k <= 8; k++)
-        L_ACF[k] = f_L_ACF[k] * scale;
+        L_ACF[k] = (int32_t)(f_L_ACF[k] * scale);
 }
 #endif /* defined (USE_FLOAT_MUL) && defined (FAST) */
