+++ src/gsm610.c
@@ -350,7 +350,7 @@ static size_t gsm610_read_f(SF_PRIVATE *psf, float *ptr, size_t len)
 		return 0;
 	pgsm610 = (GSM610_PRIVATE *)psf->codec_data;
 
-	normfact = (psf->norm_float == SF_TRUE) ? 1.0 / ((float)0x8000) : 1.0;
+	normfact = (float)((psf->norm_float == SF_TRUE) ? 1.0 / ((float)0x8000) : 1.0);
 
 	sptr = ubuf.sbuf;
 	bufferlen = ARRAY_LEN(ubuf.sbuf);
@@ -597,7 +597,7 @@ static size_t gsm610_write_f(SF_PRIVATE *psf, const float *ptr, size_t len)
 		return 0;
 	pgsm610 = (GSM610_PRIVATE *)psf->codec_data;
 
-	normfact = (psf->norm_float == SF_TRUE) ? (1.0 * 0x7FFF) : 1.0;
+	normfact = (float)((psf->norm_float == SF_TRUE) ? (1.0 * 0x7FFF) : 1.0);
 
 	sptr = ubuf.sbuf;
 	bufferlen = ARRAY_LEN(ubuf.sbuf);
@@ -605,7 +605,7 @@ static size_t gsm610_write_f(SF_PRIVATE *psf, const float *ptr, size_t len)
 	{
 		writecount = (len >= bufferlen) ? bufferlen : len;
 		for (k = 0; k < writecount; k++)
-			sptr[k] = lrintf(normfact * ptr[total + k]);
+			sptr[k] = (short)lrintf(normfact * ptr[total + k]);
 		count = gsm610_write_block(psf, pgsm610, sptr, writecount);
 
 		total += count;
@@ -635,7 +635,7 @@ static size_t gsm610_write_d(SF_PRIVATE *psf, const double *ptr, size_t len)
 	{
 		writecount = (len >= bufferlen) ? bufferlen : len;
 		for (k = 0; k < writecount; k++)
-			sptr[k] = lrint(normfact * ptr[total + k]);
+			sptr[k] = (short)lrint(normfact * ptr[total + k]);
 		count = gsm610_write_block(psf, pgsm610, sptr, writecount);
 
 		total += count;
